requirements.txt
click==8.1.7
numpy==1.26.4
opencv_python==4.9.0.80
PyYAML==6.0.1
PyYAML==6.0.1
svg.py==1.4.3
